Release 5.2.0 * CONTRIBUTING.md describes how to contribute. * Early access to the Helios cross platform camera API.   * GLES3.CameraDemo   * GLES3. DFNativeBatchCamera * Includes a small subset of the unit tests. More will be made available in future release. * Code cleanup   * FslBuildCheck now supports clang format and clang tidy.   * All code was reformatted with clang format.   * All code had a tidy pass using clang tidy. * Added OpenVX.Stereo. * Vulkan console demos no longer gets build with a window system * Vulkan.DevBatch implementation improved. * Fixed all Vulkan 1.1.85.0 validation layer warnings in:   * Vulkan.DevBatch   * Vulkan.DisplacementMapping   * Vulkan.Gears   * Vulkan.MeshInstancing   * Vulkan.TessellationPNTriangles   * Vulkan.Texturing   * Vulkan.TexturingArrays   * Vulkan.TexturingCubeMap

+++ DemoFramework/FslBase/source/FslBase/RTTI/TypeInfo.cpp
@@ -5,7 +5,9 @@ namespace Fsl
 {
   TypeInfo::TypeInfo()
   {
-    class Nil {};
+    class Nil
+    {
+    };
     m_pInfo = &typeid(Nil);
     assert(m_pInfo);
   }
@@ -22,7 +24,7 @@ namespace Fsl
   {
     assert(m_pInfo);
     // type_info::before return type is int in some VC libraries
-    return m_pInfo->before(*rhs.m_pInfo) != 0;
+    return static_cast<int>(m_pInfo->before(*rhs.m_pInfo)) != 0;
   }
